Fix folder paths
`os.path.abspath` returned a non-empty `/app` string even if `GV.ARGS.get(...)` has an empty output, leading to the function picking the first option even when it should not. This commit moves the `or` inside the function to solve this. There was no such issue for `resolve_path` at the moment, but I did the same change there for consistency and to avoid something like this in the future.

Signed-off-by: Martin Stancsics <martin.stancsics@gmail.com>